Run tests across all platform using GitHub Actions
