--- .github/workflows/test.yml
@@ -1,16 +1,0 @@
-name: Test
-
-on:
-  push:
-
-jobs:
-  build:
-    runs-on: macos-13
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Xcode 15.0
-        run: sudo xcode-select --switch /Applications/Xcode_15.0.app
-      - name: Build
-        run: swift build
-      - name: Test
-        run: swift test -v
